+++ ReportSimulationOutput/measure.rb
@@ -270,6 +270,11 @@ class ReportSimulationOutput < OpenStudio::Measure::ReportingMeasure
     arg.setDescription('Optionally generates timeseries EnergyPlus output variables. If multiple output variables are desired, use a comma-separated list. Do not include key values; by default all key values will be requested. Example: "Zone People Occupant Count, Zone People Total Heating Energy"')
     args << arg
 
+    arg = OpenStudio::Measure::OSArgument::makeStringArgument('user_output_meters', false)
+    arg.setDisplayName('Generate Timeseries Output: EnergyPlus Output Meters')
+    arg.setDescription('Optionally generates timeseries EnergyPlus output meters. If multiple output meters are desired, use a comma-separated list. Example: "Electricity:Facility, HeatingCoils:EnergyTransfer"')
+    args << arg
+
     arg = OpenStudio::Measure::OSArgument::makeStringArgument('annual_output_file_name', false)
     arg.setDisplayName('Annual Output File Name')
     arg.setDescription("If not provided, defaults to 'results_annual.csv' (or 'results_annual.json' or 'results_annual.msgpack').")
@@ -361,7 +366,7 @@ class ReportSimulationOutput < OpenStudio::Measure::ReportingMeasure
 
     args = get_arguments(runner, arguments(model), user_arguments)
 
-    setup_outputs(false, args[:user_output_variables])
+    setup_outputs(false, args)
     args = setup_timeseries_includes(@emissions, args)
 
     has_electricity_production = false
@@ -531,9 +536,14 @@ class ReportSimulationOutput < OpenStudio::Measure::ReportingMeasure
       end
     end
 
-    # Optional output variables (timeseries only)
-    @output_variables_requests.each do |output_variable_name, _output_variable|
+    # Output variables (timeseries only)
+    @output_variables_requests.each do |output_variable_name|
       result << OpenStudio::IdfObject.load("Output:Variable,*,#{output_variable_name},#{args[:timeseries_frequency]};").get
+    end
+
+    # Output meters (timeseries only)
+    @output_meters_requests.each do |output_meter_name|
+      result << OpenStudio::IdfObject.load("Output:Meter,#{output_meter_name},#{args[:timeseries_frequency]};").get
     end
 
     return result.uniq
@@ -576,7 +586,7 @@ class ReportSimulationOutput < OpenStudio::Measure::ReportingMeasure
     @hpxml_header = hpxml.header
     @hpxml_bldgs = hpxml.buildings
 
-    setup_outputs(false, args[:user_output_variables])
+    setup_outputs(false, args)
 
     if not File.exist? File.join(output_dir, 'eplusout.msgpack')
       runner.registerError('Cannot find eplusout.msgpack.')
@@ -608,7 +618,7 @@ class ReportSimulationOutput < OpenStudio::Measure::ReportingMeasure
     end
 
     if args[:timeseries_frequency] != 'none'
-      @timestamps, timestamps_dst, timestamps_utc = get_timestamps(@msgpackDataTimeseries, @hpxml_header, @hpxml_bldgs, args)
+      @timestamps, timestamps_dst, timestamps_utc = get_timestamps(@msgpackDataTimeseries, @msgpackData, @hpxml_header, @hpxml_bldgs, args)
     end
 
     # Retrieve outputs
@@ -627,15 +637,24 @@ class ReportSimulationOutput < OpenStudio::Measure::ReportingMeasure
 
   # TODO
   #
+  # @param msgpackDataTimeseries [TODO] TODO
   # @param msgpackData [TODO] TODO
   # @param hpxml_header [TODO] TODO
   # @param hpxml_bldgs [TODO] TODO
   # @param args [Hash] Map of :argument_name => value
   # @return [TODO] TODO
-  def get_timestamps(msgpackData, hpxml_header, hpxml_bldgs, args)
-    return if msgpackData.nil?
+  def get_timestamps(msgpackDataTimeseries, msgpackData, hpxml_header, hpxml_bldgs, args)
+    if not msgpackDataTimeseries.nil?
+      ep_timestamps = msgpackDataTimeseries['Rows'].map { |r| r.keys[0] }
+    elsif not msgpackData.nil?
+      msgpack_timeseries_name = get_msgpack_timeseries_name(args[:timeseries_frequency])
+      timeseries_data = msgpackData['MeterData'][msgpack_timeseries_name]
+      if not timeseries_data.nil?
+        ep_timestamps = timeseries_data['Rows'].map { |r| r.keys[0] }
+      end
+    end
 
-    ep_timestamps = msgpackData['Rows'].map { |r| r.keys[0] }
+    return if ep_timestamps.nil?
 
     if args[:add_timeseries_dst_column] || args[:use_dview_format]
       dst_start_ts = Time.utc(hpxml_header.sim_calendar_year, hpxml_bldgs[0].dst_begin_month, hpxml_bldgs[0].dst_begin_day, 2)
@@ -1190,16 +1209,36 @@ class ReportSimulationOutput < OpenStudio::Measure::ReportingMeasure
       end
     end
 
+    # Output Variables
     @output_variables = {}
-    @output_variables_requests.each do |output_variable_name, _output_variable|
+    @output_variables_requests.each do |output_variable_name|
       key_values, units = get_report_variable_data_timeseries_key_values_and_units(output_variable_name)
-      runner.registerWarning("Request for output variable '#{output_variable_name}' returned no key values.") if key_values.empty?
+      if key_values.empty?
+        runner.registerWarning("Request for output variable '#{output_variable_name}' returned no results.")
+        next
+      end
+
       key_values.each do |key_value|
         @output_variables[[output_variable_name, key_value]] = OutputVariable.new
         @output_variables[[output_variable_name, key_value]].name = "#{output_variable_name}: #{key_value.split.map(&:capitalize).join(' ')}"
         @output_variables[[output_variable_name, key_value]].timeseries_units = units
         @output_variables[[output_variable_name, key_value]].timeseries_output = get_report_variable_data_timeseries([key_value], [output_variable_name], 1, 0, args[:timeseries_frequency])
       end
+    end
+
+    # Output Meters
+    @output_meters = {}
+    @output_meters_requests.each do |output_meter_name|
+      units = get_report_meter_data_timeseries_units(output_meter_name, args[:timeseries_frequency])
+      if units.nil?
+        runner.registerWarning("Request for output meter '#{output_meter_name}' returned no results.")
+        next
+      end
+
+      @output_meters[output_meter_name] = OutputMeter.new
+      @output_meters[output_meter_name].name = output_meter_name
+      @output_meters[output_meter_name].timeseries_units = units
+      @output_meters[output_meter_name].timeseries_output = get_report_meter_data_timeseries([output_meter_name], 1, 0, args[:timeseries_frequency])
     end
 
     # Emissions
@@ -1806,17 +1845,24 @@ class ReportSimulationOutput < OpenStudio::Measure::ReportingMeasure
       output_variables_data = []
     end
 
+    # EnergyPlus output meters
+    if not @output_meters.empty?
+      output_meters_data = @output_meters.values.map { |x| [x.name, x.timeseries_units] + x.timeseries_output }
+    else
+      output_meters_data = []
+    end
+
     return if (total_energy_data.size + fuel_data.size + end_use_data.size + system_use_data.size + emissions_data.size + emission_fuel_data.size +
                emission_end_use_data.size + hot_water_use_data.size + total_loads_data.size + comp_loads_data.size + unmet_hours_data.size +
-               zone_temps_data.size + airflows_data.size + weather_data.size + resilience_data.size + output_variables_data.size) == 0
+               zone_temps_data.size + airflows_data.size + weather_data.size + resilience_data.size + output_variables_data.size + output_meters_data.size) == 0
 
     fail 'Unable to obtain timestamps.' if @timestamps.empty?
 
     if ['csv'].include? args[:output_format]
       # Assemble data
       data = data.zip(*timestamps2, *timestamps3, *total_energy_data, *fuel_data, *end_use_data, *system_use_data, *emissions_data,
-                      *emission_fuel_data, *emission_end_use_data, *hot_water_use_data, *total_loads_data, *comp_loads_data,
-                      *unmet_hours_data, *zone_temps_data, *airflows_data, *weather_data, *resilience_data, *output_variables_data)
+                      *emission_fuel_data, *emission_end_use_data, *hot_water_use_data, *total_loads_data, *comp_loads_data, *unmet_hours_data,
+                      *zone_temps_data, *airflows_data, *weather_data, *resilience_data, *output_variables_data, *output_meters_data)
 
       # Error-check
       n_elements = []
@@ -1882,7 +1928,7 @@ class ReportSimulationOutput < OpenStudio::Measure::ReportingMeasure
 
       [total_energy_data, fuel_data, end_use_data, system_use_data, emissions_data, emission_fuel_data,
        emission_end_use_data, hot_water_use_data, total_loads_data, comp_loads_data, unmet_hours_data,
-       zone_temps_data, airflows_data, weather_data, resilience_data, output_variables_data].each do |d|
+       zone_temps_data, airflows_data, weather_data, resilience_data, output_variables_data, output_meters_data].each do |d|
         d.each do |o|
           grp, name = o[0].split(':', 2)
           h[grp] = {} if h[grp].nil?
@@ -2011,10 +2057,7 @@ class ReportSimulationOutput < OpenStudio::Measure::ReportingMeasure
   def get_report_meter_data_timeseries(meter_names, unit_conv, unit_adder, timeseries_frequency)
     return [0.0] * @timestamps.size if meter_names.empty?
 
-    msgpack_timeseries_name = { 'timestep' => 'TimeStep',
-                                'hourly' => 'Hourly',
-                                'daily' => 'Daily',
-                                'monthly' => 'Monthly' }[timeseries_frequency]
+    msgpack_timeseries_name = get_msgpack_timeseries_name(timeseries_frequency)
     timeseries_data = @msgpackData['MeterData'][msgpack_timeseries_name]
     cols = timeseries_data['Cols']
     rows = timeseries_data['Rows']
@@ -2107,21 +2150,53 @@ class ReportSimulationOutput < OpenStudio::Measure::ReportingMeasure
 
   # TODO
   #
-  # @param var [TODO] TODO
+  # @param var_name [TODO] TODO
   # @return [TODO] TODO
-  def get_report_variable_data_timeseries_key_values_and_units(var)
+  def get_report_variable_data_timeseries_key_values_and_units(var_name)
     keys = []
     units = ''
-    if not @msgpackDataTimeseries.nil?
-      @msgpackDataTimeseries['Cols'].each do |col|
-        next unless col['Variable'].end_with? ":#{var}"
+    return keys, units if @msgpackDataTimeseries.nil?
 
-        keys << col['Variable'].split(':')[0..-2].join(':')
-        units = col['Units']
-      end
+    @msgpackDataTimeseries['Cols'].each do |col|
+      next unless col['Variable'].end_with? ":#{var_name}"
+
+      keys << col['Variable'].split(':')[0..-2].join(':')
+      units = col['Units']
     end
 
     return keys, units
+  end
+
+  # TODO
+  #
+  # @param meter_name [TODO] TODO
+  # @param timeseries_frequency [TODO] TODO
+  # @return [TODO] TODO
+  def get_report_meter_data_timeseries_units(meter_name, timeseries_frequency)
+    return if @msgpackData.nil?
+
+    msgpack_timeseries_name = get_msgpack_timeseries_name(timeseries_frequency)
+    timeseries_data = @msgpackData['MeterData'][msgpack_timeseries_name]
+    return if timeseries_data.nil?
+
+    timeseries_data['Cols'].each do |col|
+      next unless col['Variable'] == meter_name
+
+      return col['Units']
+    end
+
+    return
+  end
+
+  # TODO
+  #
+  # @param timeseries_frequency [TODO] TODO
+  # @return [TODO] TODO
+  def get_msgpack_timeseries_name(timeseries_frequency)
+    return { 'timestep' => 'TimeStep',
+             'hourly' => 'Hourly',
+             'daily' => 'Daily',
+             'monthly' => 'Monthly' }[timeseries_frequency]
   end
 
   # TODO
@@ -2426,11 +2501,19 @@ class ReportSimulationOutput < OpenStudio::Measure::ReportingMeasure
   end
 
   # TODO
+  class OutputMeter < BaseOutput
+    def initialize
+      super()
+    end
+    attr_accessor()
+  end
+
+  # TODO
   #
   # @param called_from_outputs_method [TODO] TODO
-  # @param user_output_variables [TODO] TODO
+  # @param args [TODO] TODO
   # @return [TODO] TODO
-  def setup_outputs(called_from_outputs_method, user_output_variables = nil)
+  def setup_outputs(called_from_outputs_method, args = {})
     # TODO
     #
     # @param fuel_type [TODO] TODO
@@ -2760,13 +2843,10 @@ class ReportSimulationOutput < OpenStudio::Measure::ReportingMeasure
     end
 
     # Output Variables
-    @output_variables_requests = {}
-    if not user_output_variables.nil?
-      output_variables = user_output_variables.split(',').map(&:strip)
-      output_variables.each do |output_variable|
-        @output_variables_requests[output_variable] = OutputVariable.new
-      end
-    end
+    @output_variables_requests = args[:user_output_variables].to_s.split(',').map(&:strip)
+
+    # Output Meters
+    @output_meters_requests = args[:user_output_meters].to_s.split(',').map(&:strip)
   end
 
   # TODO